Restructured model interfaces
